requirements.txt
Flask
Flask-SQLAlchemy
openai
Pillow
requests
